Rename infer-preferred-langauge to fit js conventions
