+++ templates/templates/base.html
@@ -13,7 +13,7 @@
   <script src="https://assets.ubuntu.com/v1/703e23c9-lazysizes+noscript+native-loading.5.1.2.min.js" defer></script>
 
   <script src="{{ versioned_static('js/dist/main.js') }}" defer></script>
-  <script src="{{ versioned_static('js/src/inferPreferredLanguage.js')}}"></script>
+  <script src="{{ versioned_static('js/src/infer-preferred-language.js')}}"></script>
 
   <link rel="stylesheet" type="text/css" media="screen" href="{{ versioned_static('css/styles.css') }}">
   <link rel="stylesheet" type="text/css" media="print" href="{{ versioned_static('css/print.css') }}">
